// src/lib/bungie/api.ts
export interface components {
  schemas: {
// ...
    "Destiny.Definitions.Records.DestinyRecordDefinition": {
      displayProperties?: components["schemas"]["Destiny.Definitions.Common.DestinyDisplayPropertiesDefinition"];
      /**
       * Format: int32
       * @description Indicates whether this Record's state is determined on a per-character or on an account-wide basis.
       */
      scope?: number;
      presentationInfo?: components["schemas"]["Destiny.Definitions.Presentation.DestinyPresentationChildBlock"];
      /** Format: uint32 */
      loreHash?: number | null;
      objectiveHashes?: number[];
      /** Format: int32 */
      recordValueStyle?: number;
      forTitleGilding?: boolean;
      /** @description A hint to show a large icon for a reward */
      shouldShowLargeIcons?: boolean;
      titleInfo?: components["schemas"]["Destiny.Definitions.Records.DestinyRecordTitleBlock"];
      completionInfo?: components["schemas"]["Destiny.Definitions.Records.DestinyRecordCompletionBlock"];
      stateInfo?: components["schemas"]["Destiny.Definitions.Records.SchemaRecordStateBlock"];
      requirements?: components["schemas"]["Destiny.Definitions.Presentation.DestinyPresentationNodeRequirementsBlock"];
      expirationInfo?: components["schemas"]["Destiny.Definitions.Records.DestinyRecordExpirationBlock"];
      /** @description Some records have multiple 'interval' objectives, and the record may be claimed at each completed interval */
      intervalInfo?: components["schemas"]["Destiny.Definitions.Records.DestinyRecordIntervalBlock"];
      /**
       * @description If there is any publicly available information about rewards earned for achieving this record, this is the list of those items.
       *  However, note that some records intentionally have "hidden" rewards. These will not be returned in this list.
       */
      rewardItems?: components["schemas"]["Destiny.DestinyItemQuantity"][];
      /** Format: int32 */
      presentationNodeType?: number;
      traitIds?: string[];
      traitHashes?: number[];
      /** @description A quick reference to presentation nodes that have this node as a child. Presentation nodes can be parented under multiple parents. */
      parentNodeHashes?: number[];
      /**
       * Format: uint32
       * @description The unique identifier for this entity. Guaranteed to be unique for the type of entity, but not globally.
       * When entities refer to each other in Destiny content, it is this hash that they are referring to.
       */
      hash?: number;
      /**
       * Format: int32
       * @description The index of the entity as it was found in the investment tables.
       */
      index?: number;
      /** @description If this is true, then there is an entity with this identifier/type combination, but BNet is not yet allowed to show it. Sorry! */
      redacted?: boolean;
    };
// ...
    "Destiny.Components.Records.DestinyRecordComponent": {
      /** Format: int32 */
      state?: number;
      objectives?: components["schemas"]["Destiny.Quests.DestinyObjectiveProgress"][];
      intervalObjectives?: components["schemas"]["Destiny.Quests.DestinyObjectiveProgress"][];
      /** Format: int32 */
      intervalsRedeemedCount?: number;
      /**
       * Format: int32
       * @description If available, this is the number of times this record has been completed. For example, the number of times a seal title has been gilded.
       */
      completedCount?: number | null;
      /** @description If available, a list that describes which reward rewards should be shown (true) or hidden (false). This property is for regular record rewards, and not for interval objective rewards. */
      rewardVisibilty?: boolean[];
    };
// ...
  };
// ...
}